Crate vec_like

This crate provides some traits that capture behaviour which some Vec-like structs have in common.

Traits

PeekBack
Get the element at the back of the list

PeekFront
Get the element at the front of the list

PopBack
Remove an element from the back of the list

PopFront
Remove an element from the front of the list

PushBack
Add an element to the back of the list

PushFront
Add an element to the front of the list