Trait vec_like::PeekBack

pub trait PeekBack<T: ?Sized> {
    fn peek_back(&mut self) -> Option<&T>;
}

Get the element at the back of the list

Required methods

fn peek_back(&mut self) -> Option<&T>

Try to get the element at the back of the list.

Loading content...