Trait vec_like::PopBack

pub trait PopBack<T> {
    fn pop_back(&mut self) -> Option<T>;
}

Remove an element from the back of the list

Required methods

fn pop_back(&mut self) -> Option<T>

Try to remove an element from the back of the list.

Loading content...