Trait vec_like::PopFront

pub trait PopFront<T> {
    fn pop_front(&mut self) -> Option<T>;
}

Remove an element from the front of the list

Required methods

fn pop_front(&mut self) -> Option<T>

Try to remove an element from the front of the list.

Loading content...