Trait vec_like::PushFront

pub trait PushFront<T> {
    type Err;
    fn push_front(&mut self, t: T) -> Result<(), Self::Err>;
}

Add an element to the front of the list

Associated Types

type Err

The type of the failure value

A sensible default is using T itself and returning the argument unmodified

Loading content...
Required methods

fn push_front(&mut self, t: T) -> Result<(), Self::Err>

Try to add an element to the front of the list.

Loading content...